[REFACTOR] Item domain 패키지 변경
