+++ side/src/main/java/com/server/side/item/domain/ItemRepository.java
@@ -1,5 +1,6 @@
-package com.server.side.item;
+package com.server.side.item.domain;
 
+import com.server.side.item.domain.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long> {
